requirements.txt
Django==4.2.5
django-cleanup==8.0.0
psycopg2-binary==2.9.7
Pillow==10.0.1
django-recaptcha==3.0.0
requests==2.31.0